refactor: video 객체 분리
